% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_SIS.R
\name{do_mut}
\alias{do_mut}
\title{draw new mutation values}
\usage{
do_mut(
  state,
  no_tradeoff,
  nt_mut_var_pos_trait,
  orig_trait,
  power_c,
  power_exp,
  mut_link_p,
  mutated,
  parasite_tuning,
  tradeoff_only,
  eff_scale,
  ...
)
}
\arguments{
\item{state}{current system state}

\item{orig_trait}{vector of original trait values}

\item{mut_var}{variable under mutation ("beta" or otherwise)}
}
\description{
draw new mutation values
}
