% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_SIS.R
\name{rbinom_mat}
\alias{rbinom_mat}
\title{Select binomial random deviates, return a matrix with appropriate dimensions}
\usage{
rbinom_mat(n, size, prob, nrow, ncol)
}
\arguments{
\item{n}{total number of deviates}

\item{size}{binomial N}

\item{prob}{binomial probability}

\item{nrow}{number of rows}

\item{ncol}{number of cols}
}
\description{
Select binomial random deviates, return a matrix with appropriate dimensions
}
