% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_SIS.R
\name{run_sim}
\alias{run_sim}
\title{Top level sim function (wrapper)}
\usage{
run_sim(
  no_tradeoff = TRUE,
  nt_mut_var_pos_trait = TRUE,
  pos_trait0 = 0.005,
  deterministic = FALSE,
  determ_mut = TRUE,
  parasite_tuning = TRUE,
  tradeoff_only = FALSE,
  agg_eff_adjust = FALSE,
  R0_init = 2,
  neg_trait0 = 0.03,
  tune0 = 0.97,
  gamma0 = 0.2,
  N = 200,
  mu = 0.01,
  mut_type = "shift",
  mut_mean = -1,
  mut_sd = 0.5,
  mut_link_p = NULL,
  power_c = 0.002,
  power_exp = 3,
  eff_hit = 1,
  Imat = NULL,
  eff_scale = 50,
  nt = 1e+05,
  rptfreq = max(nt/500, 1),
  seed = NULL,
  progress = "bar",
  debug = FALSE,
  debug2 = FALSE,
  debug3 = FALSE,
  debug4 = FALSE,
  debug4_val = 1,
  determ_length = 200,
  determ_timestep = 5,
  lsoda_hini = 0.5,
  Imat_seed = c(15, 55)
)
}
\arguments{
\item{deterministic}{FIXME}

\item{mut_type}{FIXME}

\item{Imat}{FIXME}

\item{eff_scale}{FIXME}

\item{progress}{FIXME}

\item{mut_var}{FIXME}
}
\description{
Top level sim function (wrapper)
}
