% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\docType{data}
\name{stoch_params0}
\alias{stoch_params0}
\title{default parameters}
\format{An object of class \code{list} of length 34.}
\usage{
stoch_params0
}
\description{
default parameters
}
\keyword{datasets}
