% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_SIS.R
\name{update_mut_pt}
\alias{update_mut_pt}
\title{Update mutant strain's trait values using power-law tradeoff.}
\usage{
update_mut_pt(
  state,
  orig_trait,
  power_c,
  power_exp,
  mut_link_p,
  mutated,
  mutated_host,
  no_tradeoff,
  tradeoff_only,
  ...
)
}
\description{
Update mutant strain's trait values using power-law tradeoff.
}
